Use split_inclusive() to split symlink name and target
We are about to make use of split_inclusive() soon in following patches
to parse a null terminated string and then rest of the message could
be a security context. This will be done in mknod()/mkdir()/symlink()/create()
paths.

So this patch prepares the ground and moves symlink() to making use of
split_inclusive(). It needs to parse two null terminated strings already.

No functional change. This will just make it easier to use this code
in later patches when security context is being parsed.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>