Add an helper to set security context on mkdir/mknod/symlink
Add an helper which can be used to set security context on mkdir/mknod/symlink
calls.

create() is different. We get an open fd and it creates regular files so
we can simply call fsetxattr().

With mknodat()/mkdir()/symlink() calls, we don't get an open fd. We will
have to open one explicitly. For device nodes it is not recommended to
open a device node on host and it can have side affects. So instead of
opening node, we open an O_PATH descriptor and then use setxattr() on
proc path to make it safe for device nodes.

For now use same logic for mkdir() and symlink() as well. Though I think
we can use fsetxattr() atleast for directory creation. We can optimize it
some other time. Keep it simple for now.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>