Bump vhost-user-backend and virtio-queue dependencies
Updating dependencies to:
- virtio-queue 0.5.0
- vhost-user-backend 0.6.0

This latest virtio-queue version introduces a few breaking changes,
which is why this patch slightly updates the code.

The latest vhost-user-backend version relies on the same virtio-queue
version, but it also contains a fix for proper vring initialization so
that booting a VM with OVMF firmware works as expected.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>