Fix `listxattr()` response when xattr name list is empty
FUSE_LISTXATTR returns a xattr name list. It is possible that the name
list is empty, because either there are no xattrs or all xattrs have
been filtered out by the server.

As of now rust version of code writes a null byte to client supplied
buffer. This fails the kernel fuse code verification and return -EIO to
user space.

We need to send a response back with error code 0 and virtiofs kernel
code will handle it and update the variable argument size to 0.

Signed-off-by: German Maglione <gmaglione@redhat.com>