sandbox: Allow non-root users run the daemon without isolation
Although, virtiofsd expect to run as root, either as the real one,
or the pseudo-root inside a user namespace, there are use cases
there are use cases that need to be able to run virtiofsd as a
non-privileged user outside a user namespace.

In this mode many operations will not be available (e.g. chown).

Signed-off-by: German Maglione <gmaglione@redhat.com>