virtiofsd: add option "--preserve-atime"
When running in unprivileged mode and with `--sandbox=none` we
automatically clean up O_NOATIME to prevent potential permission errors
on the exported files/directories, since virtiofsd may not have full
ownership/permissions on them.

However, sometimes we may want to preserve O_NOATIME without implicitly
cleaning it.

For this reason provide the option `--preserve-atime` to not modify the
O_NOATIME behavior of the client.

Signed-off-by: Andrea Righi <andrea.righi@canonical.com>