Fix cache timeouts
Currently, we set both the entry and the attr timeout to 5s, we should
set a proper value for each cache policy. For instance, it doesn't make
sense to have a timeout greater than 0 if the cache is disabled, since
the guest kernel must send a lookup request on each attempt to look up
an entry.

Signed-off-by: German Maglione <gmaglione@redhat.com>