Allow to provide the same argument multiple times
By default, clap  does not allow to provide the same argument more than
once, for instance if we call:
`virtiofsd --xattr --xattr ...`

it will show the following error:
`error: The argument '--xattr' was provided more than once, but cannot
be used multiple times`

This is not compatible with the C version, and also it breaks the use of
annotations in Kata:
`io.katacontainers.config.hypervisor.virtio_fs_extra_args`
it appends the extra arguments to the default ones used by Kata,
failing to start the pod if a duplicated argument is provided, for
instance `--thread-pool-size=`.

So, let's add `AppSettings::AllArgsOverrideSelf` to the global
settings, essentially this sets `overrides_with("<iself>")` for each
argument. If we need a finer control, we can remove the global setting
and add `overrides_with` only to arguments that are allowed to be
duplicated, for instance:
```
    #[structopt(long, overrides_with("xattr"))]
    xattr: bool,
```

Signed-off-by: German Maglione <gmaglione@redhat.com>