Set the version after the release as "dirty"
Some projects use the latest upstream version of virtiofsd (the last
commit). And keeping the latest release version makes it difficult to
troubleshoot problems.

A proper way to do this should be with something like build.rs +
git describe, but since compilation dependencies would be added,
let's set the version of the upstream as "dirty" for now.

Signed-off-by: German Maglione <gmaglione@redhat.com>