Use `name_to_handle_at()` to get the mount id
Only since linux 5.8, `statx()` returns the mount ID of the mount
containing the file passed as parameter (via `STATX_MNT_ID` in the
`stx_mask` field). Currently, if `statx()` does not return the mount
id, it defaults to `0`.

This in combination with that we use the tuple (inode no., device,
mount id) as the key in the `InodeStore`, in kernels prior to 5.8,
sometimes makes the wrong `inode` to be retrieved from the
`InodeStore`.

For instance, let's assume a kernel < 5.8 (i.e., `mnt_id` is == 0)
and this filesystem structure:
```
#!/bin/bash

kroot=bind-test
mkdir -p ${kroot}/{data,share}
cd ${kroot}

mkdir -p share/{data-ro,data-rw}

mount -o bind,ro  data share/data-ro
mount -o bind,rw  data share/data-rw
```

where both `share/data-ro` and `share/data-rw` are bind mounts of
`data` (read-only and read-write respectively). When looking at
`/proc/self/mountinfo` the only difference between those two is
the mount id.

When sharing `share` and trying to create a file in `data-ro` we will
get the "read-only file system error", but then when trying to write to
`data-rw` we will get the same error because we cannot differentiate
between `data-ro` and `data-rw` because both have the same key (since
`mnt_id == 0`). And on the opposite, if we try to write first in
`data-rw` and then in `data-ro` both writes will be successful,
basically the first one that enters the `InodeStore` is the one that
will be used in all cases.

This was reported on the Kata Containers channel by Eric Ernst, while
trying to start a pod with 2 containers, each having a rw and a ro
mount of a shared emptyDir.

So before returning `0` in case `statx()` does not return the mount id,
let's try to use `name_to_handle_at()`, since it does not require
special permissions and most file systems support file handles.

Signed-off-by: German Maglione <gmaglione@redhat.com>