Add support for VHOST_USER_PROTOCOL_F_CONFIG
This adds --tag parameter and advertises this tag via the vhost-user
F_CONFIG feature.

While the spec does not seem to disallow zero-length strings as tag,
we will only actually advertise F_CONFIG if --tag was specified. This
preserves backwards-compatibility, especially when considering the QEMU
situation around F_CONFIG support.

Background:

Traditionally, vhost-user frontends (read as: QEMU) cared about
providing the CONFIG space. F_CONFIG is a flag that indicates that the
backend can provide the configuration space on its own. While this - in
theory - allows simpler frontend implementations, the reality is that
most vhost-user backends just ended up implicitly depending on QEMU to
advertise it.

In fact, QEMU is only querying the CONFIG section of the backend for
very few device classes. virtio-fs devices are currently not among them.
Even worse, a patch [1] that intended to at least make the user aware of
the situation, accidentally crashed QEMU while printing the warning.

This leaves F_CONFIG support in a ugly situation for QEMU. But,
F_CONFIG is standardized since a while and we are working on improving
the QEMU situation [2]. A world where the backends provide their
own configuration simplifies the needs on the frontend and makes the
backends more reusable across hypervisors at the same time.

New frontends - like the Xen vhost-user frontend - depend on this to
avoid needing to fulfill the historical expectations that evolved around
QEMU's vhost-user daemons. Once the QEMU situation is improved, this
change will also become useful for QEMU users.

Additionally - with F_CONFIG existing - there is little point in
requiring every other frontend implementation to cover all the quirks
that had to evolve in QEMU over the time.

[1] https://gitlab.com/qemu-project/qemu/-/commit/56534930
[2] https://lore.kernel.org/r/20230710153522.3469097-1-alex.bennee@linaro.org/

Signed-off-by: Erik Schilling <erik.schilling@linaro.org>