README: Use memfd for qemu by default
We run qemu and virtiofs from a Kubernetes container, and the
default for `/dev/shm` for historical reasons is 64MiB, which
is insufficient for most virtual machines.

There's no reason to persist RAM to a filesystem by default; I believe
it causes suboptimal performance.

Instead, allocate a memfd, which is independent of any particular
filesystem and thus inherits the ambient memory limits (e.g. cgroups).

Signed-off-by: Colin Walters <walters@verbum.org>