README: Add guidance for adding to qemu cmdlines
If you're not fully familiar with all of the details of running qemu
then you might see a relatively simple looking qemu command-line
and assume just adding the new bits will make it work.

If you miss that the memory size of the memfd option matches
the memory of the VM then you might start with incompatible options
which can hang without explanation why.

This patch adds a brief explanation of the qemu commands
that need to be added or updated in order to add virtiofs devices,
and a reference from the example.

Signed-off-by: Richard Maw <richard.maw@codethink.co.uk>